// packages/spreadsheet-component/src/types/xlsx.ts
export interface XLSXStructure {
  relsFiles: XLSXRelFile[];
  sharedStrings: string[];
  fonts: XLSXFont[];
  fills: XLSXFill[];
  borders: XLSXBorder[];
  numFmts: XLSXNumFormat[];
  styles: XLSXStyle[];
  dxfs: XLSXDxf[];
}

export interface XLSXImportData extends Omit<XLSXStructure, "relsFiles"> {
  sheets: XLSXWorksheet[];
  externalBooks: XLSXExternalBook[];
}
// ...
export interface XLSXDxf {
  font?: XLSXFont;
  fill?: XLSXFill;
  numFmt?: XLSXNumFormat;
  alignment?: XLSXCellAlignment;
  border?: XLSXBorder;
}

export interface XLSXRel {
  id: string;
  type: string;
  target: string;
  targetMode?: "External";
}

export interface XLSXRelFile {
  path: string;
  rels: XLSXRel[];
}

export type XLSXExportFile = XLSXExportImageFile | XLSXExportXMLFile;

export interface XLSXExportXMLFile {
  path: string;
  content: string;
  contentType?: string;
}

export interface XLSXExportImageFile {
  path: string;
  imageSrc: string;
}

export interface XLSXExport {
  name: string;
  files: XLSXExportFile[];
}

export interface XLSXFont {
  size?: number;
  family?: number;
  color?: XLSXColor;
  name?: string;
  bold?: boolean;
  italic?: boolean;
  underline?: boolean;
  strike?: boolean;
}
export interface XLSXCellAlignment {
  horizontal?: XLSXHorizontalAlignment;
  vertical?: XLSXVerticalAlignment;
  textRotation?: number;
  wrapText?: boolean;
  indent?: number;
  relativeIndent?: number;
  justifyLastLine?: boolean;
  shrinkToFit?: boolean;
  readingOrder?: number;
}

export interface XLSXFill {
  patternType?: XLSXFillPatternType;
  reservedAttribute?: string; // will generate a specific specific attribute in XML. If set, fgColor is ignored.
  fgColor?: XLSXColor;
  bgColor?: XLSXColor;
}

export interface XLSXStyle {
  fontId: number;
  fillId: number;
  borderId: number;
  numFmtId: number;
  alignment?: XLSXCellAlignment;
}

export interface ExtractedStyle {
  font: XLSXFont;
  fill: XLSXFill;
  border: number;
  numFmt: XLSXNumFormat | undefined;
  alignment: XLSXCellAlignment;
}

export interface XLSXWorksheet {
  sheetName: string;
  isVisible: boolean;
  sheetViews: XLSXSheetView[];
  sheetFormat?: XLSXSheetFormat;
  sheetProperties?: XLSXSheetProperties;
  cols: XLSXColumn[];
  rows: XLSXRow[];
  cfs: XLSXConditionalFormat[];
  sharedFormulas: string[];
  merges: string[];
  figures: XLSXFigure[];
  hyperlinks: XLSXHyperLink[];
  tables: XLSXTable[];
  pivotTables: XLSXPivotTable[];
}

export interface XLSXSheetView {
  /** True if this is the sheet currently selected */
  tabSelected: boolean;
  showFormulas: boolean;
  showGridLines: boolean;
  showRowColHeaders: boolean;
  pane: PaneDivision;
}

export interface XLSXSheetFormat {
  defaultColWidth: number;
  defaultRowHeight: number;
}

export interface XLSXColumn {
  min: number;
  max: number;
  width?: number;
  customWidth?: boolean;
  bestFit?: boolean;
  hidden?: boolean;
  styleIndex?: number;
  outlineLevel?: number;
  collapsed?: boolean;
}
export interface XLSXRow {
  index: number;
  height?: number;
  customHeight?: boolean;
  hidden?: boolean;
  cells: XLSXCell[];
  styleIndex?: number;
  outlineLevel?: number;
  collapsed?: boolean;
}

export interface XLSXFormula {
  content?: string;
  sharedIndex?: number;
  ref?: string;
}

export interface XLSXCell {
  xc: string; // OpenXml specs defines it as optional, but not having it makes no sense
  styleIndex?: number;
  type: XLSXCellType;
  value?: string;
  formula?: XLSXFormula;
}
export interface XLSXTheme {
  clrScheme?: XLSXColorScheme[];
}

export interface XLSXColorScheme {
  name: string;
  value: string;
  lastClr?: string;
}

export interface XLSXNumFormat {
  id: number;
  format: Format;
}
export interface XLSXBorder {
  top?: XLSXBorderDescr;
  left?: XLSXBorderDescr;
  bottom?: XLSXBorderDescr;
  right?: XLSXBorderDescr;
  diagonal?: XLSXBorderDescr;
  diagonalUp?: boolean;
  diagonalDown?: boolean;
}
// ...
export interface XLSXConditionalFormat {
  cfRules: XLSXCfRule[];
  sqref: string[];
  pivot?: boolean;
}

export interface XLSXCfRule {
  type: XLSXCfType;
  priority: number;
  formula?: string[];
  colorScale?: XLSXColorScale;
  dataBar?: any;
  iconSet?: XLSXIconSet;
  dxfId?: number;
  stopIfTrue?: boolean;
  aboveAverage?: boolean;
  percent?: boolean;
  bottom?: boolean;
  operator?: XLSXCfOperatorType;
  text?: string;
  timePeriod?: any;
  rank?: number;
  stdDev?: number;
  equalAverage?: boolean;
}

export interface XLSXSharedFormula {
  formula: string;
  refCellXc: string;
}

export interface XLSXColor {
  auto?: boolean;
  indexed?: number;
  rgb?: string;
  tint?: number;
}

export interface XLSXFigureAnchor {
  col: number;
  colOffset: number; // in EMU (English Metrical Unit)
  row: number;
  rowOffset: number; // in EMU (English Metrical Unit)
}

export interface XLSXFigureSize {
  cx: number;
  cy: number;
}

export interface XLSXFigure {
  anchors: XLSXFigureAnchor[];
  data: ExcelChartDefinition | ExcelImage;
  figureSize?: ExcelFigureSize;
}
// ...
export interface XLSXHyperLink {
  xc: string;
  location?: string;
  display?: string;
  relTarget?: string;
}

export interface XLSXTableStyleInfo {
  name?: string;
  showFirstColumn?: boolean;
  showLastColumn?: boolean;
  showRowStripes?: boolean;
  showColumnStripes?: boolean;
}

export interface XLSXPivotTable {
  name: string;
  rowGrandTotals: boolean;
  location: XLSXPivotTableLocation;
  style?: XLSXPivotTableStyleInfo;
}

export interface XLSXPivotTableLocation {
  ref: string;
  firstHeaderRow: number;
  firstDataRow: number;
  firstDataCol: number;
}

export interface XLSXPivotTableStyleInfo {
  name: string;
  showRowHeaders: boolean;
  showColHeaders: boolean;
  showRowStripes: boolean;
  showColStripes: boolean;
  showLastColumn?: boolean;
}

export interface XLSXTableCol {
  name: string;
  id: string;
  colFormula?: string;
}

export interface XLSXTable {
  displayName: string;
  name?: string;
  id: string;
  ref: string;
  headerRowCount: number;
  totalsRowCount: number;
  cols: XLSXTableCol[];
  style?: XLSXTableStyleInfo;
  autoFilter?: XLSXAutoFilter;
}

export interface XLSXAutoFilter {
  zone: string;
  columns: XLSXFilterColumn[];
}

export interface XLSXFilterColumn {
  colId: number;
  hiddenButton?: boolean;
  filters: XLSXSimpleFilter[];
}

export interface XLSXSimpleFilter {
  val: string;
}

export interface XLSXExternalBook {
  rId: string;
  sheetNames: string[];
  datasets: XLSXExternalSheetData[];
}

export interface XLSXExternalSheetData {
  sheetId: number;
  data: Record<string, string>; // Record XC : value
}

export type XLSXSheetState = "visible" | "hidden" | "veryHidden";

export interface XLSXSheetWorkbookInfo {
  relationshipId: string;
  sheetId: string;
  sheetName: string;
  state: XLSXSheetState;
}

export interface XLSXSheetProperties {
  outlinePr?: XLSXOutlineProperties;
  tabColor?: XLSXColor;
}

export interface XLSXOutlineProperties {
  summaryBelow: boolean;
  summaryRight: boolean;
}